+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/ScenarioParametersApiHttpServiceSecuritySpec.scala
@@ -107,6 +107,41 @@ class ScenarioParametersApiHttpServiceSecuritySpec
           )
       }
     }
+    "impersonating user has permission to impersonate should" - {
+      "return parameters combination for categories the impersonated user has a write access" in {
+        given()
+          .when()
+          .basicAuthAllPermUser()
+          .impersonateLimitedWriterUser()
+          .get(s"$nuDesignerHttpAddress/api/scenarioParametersCombinations")
+          .Then()
+          .statusCode(200)
+          .equalsJsonBody(
+            s"""{
+               |  "combinations": [
+               |    {
+               |      "processingMode": "Unbounded-Stream",
+               |      "category": "Category1",
+               |      "engineSetupName": "Flink"
+               |    }
+               |  ],
+               |  "engineSetupErrors": {}
+               |}""".stripMargin
+          )
+      }
+    }
+    "impersonating user does not have permission to impersonate should" - {
+      "forbid access" in {
+        given()
+          .when()
+          .basicAuthWriter()
+          .impersonateLimitedWriterUser()
+          .get(s"$nuDesignerHttpAddress/api/scenarioParametersCombinations")
+          .Then()
+          .statusCode(403)
+          .equalsPlainBody("The supplied authentication is not authorized to impersonate")
+      }
+    }
   }
 
 }