+++ security/src/main/scala/pl/touk/nussknacker/ui/security/api/LoggedUser.scala
@@ -2,23 +2,89 @@ package pl.touk.nussknacker.ui.security.api
 
 import pl.touk.nussknacker.ui.security.api.AuthenticationConfiguration.ConfigRule
 import GlobalPermission.GlobalPermission
+import pl.touk.nussknacker.security.Permission
 import pl.touk.nussknacker.security.Permission.Permission
+import pl.touk.nussknacker.ui.security.api.CreationError.ImpersonationNotAllowed
 
 sealed trait LoggedUser {
   val id: String
   val username: String
-  val isAdmin: Boolean
   def can(category: String, permission: Permission): Boolean
 }
 
 object LoggedUser {
 
-  def apply(
+  def create(
       authenticatedUser: AuthenticatedUser,
-      rules: List[ConfigRule]
-  ): LoggedUser = {
+      rules: List[ConfigRule],
+      isAdminImpersonationPossible: Boolean = false
+  ): Either[CreationError, LoggedUser] = {
+    val loggedUser = RealLoggedUser(authenticatedUser, rules)
+    authenticatedUser.impersonatedAuthenticationUser match {
+      case None =>
+        Right(loggedUser)
+      case Some(impersonatedUser) if loggedUser.canImpersonate =>
+        createImpersonatedUser(
+          loggedUser,
+          RealLoggedUser(impersonatedUser, rules),
+          isAdminImpersonationPossible
+        )
+      case Some(_) =>
+        Left(ImpersonationNotAllowed)
+    }
+  }
+
+  private def createImpersonatedUser(
+      impersonatingUser: RealLoggedUser,
+      impersonatedLoggedUser: RealLoggedUser,
+      isAdminImpersonationPossible: Boolean
+  ): Either[CreationError, LoggedUser] = impersonatedLoggedUser match {
+    case _: CommonUser =>
+      Right(ImpersonatedUser(impersonatedLoggedUser, impersonatingUser))
+    case _: AdminUser if isAdminImpersonationPossible =>
+      Right(ImpersonatedUser(impersonatedLoggedUser, impersonatingUser))
+    case _: AdminUser =>
+      Left(ImpersonationNotAllowed)
+  }
+
+  implicit class UserImpersonationSyntax(val user: LoggedUser) extends AnyVal {
+
+    def canImpersonate: Boolean = user match {
+      case _: AdminUser        => true
+      case _: ImpersonatedUser => false
+      case commonUser: CommonUser =>
+        commonUser.globalPermissions
+          .map(_.toLowerCase)
+          .contains(Permission.Impersonate.toString.toLowerCase)
+    }
+
+    def isAdmin: Boolean = user match {
+      case _: AdminUser        => true
+      case _: CommonUser       => false
+      case u: ImpersonatedUser => u.impersonatedUser.isAdmin
+    }
+
+    def impersonatingUserId: Option[String] = user match {
+      case _: RealLoggedUser   => None
+      case u: ImpersonatedUser => Some(u.impersonatingUser.id)
+    }
+
+    def impersonatingUserName: Option[String] = user match {
+      case _: RealLoggedUser   => None
+      case u: ImpersonatedUser => Some(u.impersonatingUser.username)
+    }
+
+  }
+
+}
+
+sealed trait RealLoggedUser extends LoggedUser
+
+object RealLoggedUser {
+
+  def apply(authenticatedUser: AuthenticatedUser, rules: List[ConfigRule]): RealLoggedUser = {
     val rulesSet = RulesSet.getOnlyMatchingRules(authenticatedUser.roles.toList, rules)
-    apply(id = authenticatedUser.id, username = authenticatedUser.username, rulesSet = rulesSet)
+    RealLoggedUser(id = authenticatedUser.id, username = authenticatedUser.username, rulesSet = rulesSet)
   }
 
   def apply(
@@ -27,7 +93,7 @@ object LoggedUser {
       categoryPermissions: Map[String, Set[Permission]] = Map.empty,
       globalPermissions: List[GlobalPermission] = Nil,
       isAdmin: Boolean = false
-  ): LoggedUser = {
+  ): RealLoggedUser = {
     if (isAdmin) {
       AdminUser(id, username)
     } else {
@@ -35,11 +101,11 @@ object LoggedUser {
     }
   }
 
-  def apply(id: String, username: String, rulesSet: RulesSet): LoggedUser = {
+  private def apply(id: String, username: String, rulesSet: RulesSet): RealLoggedUser = {
     if (rulesSet.isAdmin) {
-      LoggedUser(id = id, username = username, isAdmin = true)
+      RealLoggedUser(id = id, username = username, isAdmin = true)
     } else {
-      LoggedUser(
+      RealLoggedUser(
         id = id,
         username = username,
         categoryPermissions = rulesSet.permissions,
@@ -55,20 +121,31 @@ final case class CommonUser(
     username: String,
     categoryPermissions: Map[String, Set[Permission]] = Map.empty,
     globalPermissions: List[GlobalPermission] = Nil
-) extends LoggedUser {
-
-  def categories(permission: Permission): Set[String] = categoryPermissions.collect {
-    case (category, permissions) if permissions contains permission => category
-  }.toSet
+) extends RealLoggedUser {
 
   override def can(category: String, permission: Permission): Boolean = {
     categoryPermissions.get(category).exists(_ contains permission)
   }
 
-  override val isAdmin: Boolean = false
+  def categories(permission: Permission): Set[String] = categoryPermissions.collect {
+    case (category, permissions) if permissions contains permission => category
+  }.toSet
+
 }
 
-final case class AdminUser(id: String, username: String) extends LoggedUser {
+final case class AdminUser(id: String, username: String) extends RealLoggedUser {
   override def can(category: String, permission: Permission): Boolean = true
-  override val isAdmin: Boolean                                       = true
+}
+
+final case class ImpersonatedUser(impersonatedUser: RealLoggedUser, impersonatingUser: RealLoggedUser)
+    extends LoggedUser {
+  override val id: String                                             = impersonatedUser.id
+  override val username: String                                       = impersonatedUser.username
+  override def can(category: String, permission: Permission): Boolean = impersonatedUser.can(category, permission)
+}
+
+sealed trait CreationError
+
+object CreationError {
+  case object ImpersonationNotAllowed extends CreationError
 }